[Backend] Move target helpers under backend
