+++ examples/gemm_fp8/example_tilelang_gemm_amd_fp8_preshuffle.py
@@ -7,7 +7,7 @@ import tilelang.language as T
 from tilelang.tileop.base import GemmWarpPolicy
 from tilelang.layout import make_swizzled_layout
 from tilelang.rocm.intrinsics.mfma_macro_generator import MatrixCorePreshuffleIntrinEmitter
-from tilelang.utils import determine_fp8_type
+from tilelang.backend.target import determine_fp8_type
 
 tilelang.testing.set_random_seed(0)
