+++ testing/python/debug/test_tilelang_debug_print.py
@@ -3,7 +3,7 @@ import pytest
 import tilelang
 import tilelang.testing
 import tilelang.language as T
-from tilelang.utils import determine_fp8_type
+from tilelang.backend.target import determine_fp8_type
 
 
 def debug_print_buffer(M=16, N=16, dtype=T.float16):
