+++ testing/python/transform/test_tilelang_transform_lower_hopper_intrin.py
@@ -1,7 +1,7 @@
 from tilelang import tvm as tvm
 import tilelang as tl
 from tilelang.cuda import transform as cuda_transform
-from tilelang.utils.target import determine_target
+from tilelang.backend.target import determine_target
 import tilelang.language as T
 import tilelang.testing
