+++ testing/python/transform/test_tilelang_transform_lower_shared_barrier.py
@@ -1,7 +1,7 @@
 # ruff: noqa
 from tilelang import tvm as tvm
 import tilelang as tl
-from tilelang.utils.target import determine_target
+from tilelang.backend.target import determine_target
 import tilelang.language as T
 import tilelang.testing
 from tilelang.cuda.pipeline import CUDAPassPipelineBodyPrologue
