+++ testing/python/transform/test_tilelang_transform_pipeline_planning.py
@@ -1,6 +1,6 @@
 from tilelang import tvm as tvm
 import tilelang as tl
-from tilelang.utils.target import determine_target
+from tilelang.backend.target import determine_target
 import tilelang.language as T
 from tvm.tirx.stmt_functor import post_order_visit
