+++ tilelang/autotuner/param.py
@@ -474,7 +474,7 @@ class AutotuneResult:
 
         verbose = compile_args.verbose
         # Normalize target and resolve execution backend for loading
-        from tilelang.utils.target import determine_target as _determine_target
+        from tilelang.backend.target import determine_target as _determine_target
         from tilelang.backend.execution_backend import resolve_execution_backend
 
         norm_target = _determine_target(compile_args.target, return_object=True)
