+++ tilelang/carver/arch/metal.py
@@ -10,7 +10,7 @@ def is_metal_arch(arch: TileDevice) -> bool:
 class METAL(TileDevice):
     def __init__(self, target: Target | str):
         if isinstance(target, str):
-            from tilelang.utils.target import determine_target
+            from tilelang.backend.target import determine_target
 
             target = determine_target(target, return_object=True)
         self.target = target
