+++ tilelang/engine/lower.py
@@ -15,7 +15,7 @@ from tilelang.env import COMPOSABLE_KERNEL_INCLUDE_DIR, CUTLASS_INCLUDE_DIR, TIL
 from tilelang.transform import PassConfigKey
 from tilelang.engine.param import KernelParam, CompiledArtifact
 from tilelang.engine.semantic_check import PreLowerSemanticCheck
-from tilelang.utils.target import determine_target, target_get_mcpu
+from tilelang.backend.target import determine_target, target_get_mcpu
 from tilelang.backend.pass_pipeline import resolve_pipeline
