+++ tilelang/jit/adapter/cython/adapter.py
@@ -17,7 +17,7 @@ from tilelang.jit.adapter.base import BaseKernelAdapter, CachedTextSource
 from tilelang.jit.adapter.wrapper import TLWrapper
 from tilelang.jit.adapter.libgen import LibraryGenerator
 from tilelang.jit.adapter.utils import is_cuda_target, is_hip_target, is_cpu_target, is_metal_target
-from tilelang.utils.target import determine_target
+from tilelang.backend.target import determine_target
 from tilelang.utils.language import retrieve_func_from_module
 
 logger = logging.getLogger(__name__)
