+++ tilelang/metal/op/gemm/__init__.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from tilelang.tileop.gemm.registry import register_gemm_impl
-from tilelang.utils.target import target_is_metal
+from tilelang.backend.target import target_is_metal
 from .gemm_metal import GEMM_INST_METAL, GemmMetal
