+++ tilelang/utils/__init__.py
@@ -1,10 +1,5 @@
 """The profiler and convert to torch utils"""
 
-from .target import (  # noqa: F401
-    determine_target,
-    determine_fp8_type,
-    determine_torch_fp8_type,
-)
 from .tensor import TensorSupplyType, torch_assert_close  # noqa: F401
 from .language import (
     is_global,  # noqa: F401
